rename updateInfoFormHandler
