+++ ft_transcendence/static/js/app/handlers/updateInfoFormHandler.js
